+++ mz_strm_wzaes.c
@@ -215,7 +215,7 @@ int32_t mz_stream_wzaes_read(void *stream, void *buf, int32_t size)
     int32_t read = 0;
 
     max_total_in = wzaes->max_total_in - MZ_AES_FOOTER_SIZE;
-    if (bytes_to_read > (int32_t)(max_total_in - wzaes->total_in))
+    if ((int64_t)bytes_to_read > (max_total_in - wzaes->total_in))
         bytes_to_read = (int32_t)(max_total_in - wzaes->total_in);
 
     read = mz_stream_read(wzaes->stream.base, buf, bytes_to_read);
